+++ src/components/sales/Sales.tsx
@@ -1,34 +1,15 @@
 import useSalesStore from '@/shared/store/sales';
-import clsx from 'clsx';
 import SalesDeatilWithCalendar from './salesCalendar/SalesDetailWithCalendar';
 import Status from './status/Status';
-import Tab from './status/tab/Tab';
 import styles from './styles/sales.module.css';
+import Toggle from './toggle/Toggle';
 
 const Sales = () => {
-  const { isChangeView, setIsChangeView } = useSalesStore();
+  const isChangeView = useSalesStore(state => state.isChangeView);
 
-  const clickShowChart = () => setIsChangeView(true);
-  const clickShowCalendar = () => setIsChangeView(false);
   return (
     <div className={styles.salesContainer}>
-      <div className={styles.toggleWrapper}>
-        <Tab />
-        <div className={styles.toggleBtnWrapper}>
-          <button
-            className={clsx(styles.toggleButton, { [styles.clickedToggle]: isChangeView })}
-            onClick={clickShowChart}
-          >
-            매출 관리
-          </button>
-          <button
-            className={clsx(styles.toggleButton, { [styles.clickedToggle]: !isChangeView })}
-            onClick={clickShowCalendar}
-          >
-            매출 달력
-          </button>
-        </div>
-      </div>
+      <Toggle />
       {isChangeView ? <Status /> : <SalesDeatilWithCalendar />}
     </div>
   );
