+++ src/components/sales/calendar/styles/calendar.module.css
@@ -12,28 +12,24 @@
   background: #fff;
 }
 
-.hidden-calendar {
-  display: none;
-}
-
-/*ui 개선*/
-.show-calendar {
-  margin-top: 10rem;
-  max-width: 152rem;
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  padding-bottom: 5rem;
-}
-.calendar-header-wrapper {
-  /**/
-}
 .calendar-body-wrapper {
   display: flex;
   flex-direction: column;
   align-items: flex-start;
   gap: 2rem;
+  max-width: 152rem;
+  margin: 0 auto;
   padding: 3rem;
   border-radius: 2rem;
   border: 1px solid var(--line, #ededed);
+}
+/*sales/calendar 일 때*/
+.show-calendar {
+  display: flex;
+  align-items: flex-start;
+  margin-top: 2rem;
+  padding-bottom: 5rem;
+}
+.calendar-header-wrapper {
+  /**/
 }
