+++ src/components/sales/styles/sales.module.css
@@ -11,35 +11,3 @@ scroll을 숨겼습니다.
 .sales-container::-webkit-scrollbar {
   display: none;
 }
-
-.toggle-wrapper {
-  display: flex;
-  justify-content: space-between;
-}
-
-.toggle-btn-wrapper {
-  display: flex;
-  align-items: flex-start;
-  gap: 1rem;
-}
-
-.toggle-button {
-  display: flex;
-  padding: 1rem 4rem;
-  align-items: flex-start;
-  gap: 1rem;
-  border-radius: 0.5rem;
-  background: var(--bg-g);
-  cursor: pointer;
-}
-.clickedToggle {
-  border-radius: 0.5rem;
-  background: var(--point1-80);
-  color: var(--text5);
-  text-align: center;
-  font-family: Pretendard;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 1.8rem; /* 100% */
-  letter-spacing: -0.036rem;
-}
