+++ src/components/sales/toggle/styles/toggle.module.css
@@ -1,0 +1,4 @@
+.toggle-wrapper {
+  display: flex;
+  justify-content: space-between;
+}
