+++ src/components/sales/toggle/tab/tabButton/TabButton.tsx
@@ -1,0 +1,5 @@
+const TabButton = () => {
+  return <div></div>;
+};
+
+export default TabButton;
