+++ src/components/sales/toggle/toggleButton/styles/toggleButton.module.css
@@ -1,0 +1,27 @@
+.toggle-btn-wrapper {
+  display: flex;
+  align-items: flex-start;
+  gap: 1rem;
+}
+
+.toggle-button {
+  display: flex;
+  padding: 1rem 4rem;
+  align-items: flex-start;
+  gap: 1rem;
+  border-radius: 0.5rem;
+  color: var(--text2);
+  background: var(--bg-g);
+  cursor: pointer;
+}
+.clickedToggle {
+  border-radius: 0.5rem;
+  background: var(--point1-80);
+  color: var(--text5);
+  text-align: center;
+  font-family: Pretendard;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 1.8rem; /* 100% */
+  letter-spacing: -0.036rem;
+}
